🎨 [style] 1. UI 제작 : Home, Detail, Profile storyboard 분할 구현 2. VC제작 : Home, Detail, Profile .swift파일 분할 구현 3. MVC패턴으로 폴더링
