+++ DiaryApp/DiaryApp/Controllers/Detail/DetailPostEditViewController.swift
@@ -1,0 +1,29 @@
+//
+//  DetailPostEditViewController.swift
+//  DiaryApp
+//
+//  Created by 모상현 on 2022/10/12.
+//
+
+import UIKit
+
+class DetailPostEditViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
