+++ DiaryApp/DiaryApp/SceneDelegate.swift
@@ -13,7 +13,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
 
 	func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        window!.rootViewController = UIStoryboard(name: "Home", bundle: nil).instantiateInitialViewController()!
+        window?.rootViewController = UIStoryboard(name: "Home", bundle: nil).instantiateInitialViewController()!
 		// Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
 		// If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
 		// This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
